changed shuffle to use 'regular' ints instead of NS ints

+++ GoFishModel/Deck.m
@@ -56,11 +56,11 @@
 
 - (void)shuffle {
     {
-        NSUInteger count = [[self numberOfCards] intValue];
-        for (NSUInteger i = 0; i < count; ++i) {
+        int count = [[self numberOfCards] intValue];
+        for (int i = 0; i < count; ++i) {
             // Select a random element between i and end of array to swap with.
-            NSInteger nElements = count - i;
-            NSInteger n = arc4random_uniform(nElements) + i;
+            int nElements = count - i;
+            int n = arc4random_uniform(nElements) + i;
             [self.cards exchangeObjectAtIndex:i withObjectAtIndex:n];
         }
     }
